+++ lib/caching/index.ts
@@ -1,3 +1,4 @@
+import {LRUCache} from './LRU';
 import {memoize} from './memoize';
 
-export {memoize};
+export {LRUCache, LRUCache as LRU, memoize};