+++ lib/caching/memoize.ts
@@ -1,7 +1,6 @@
 import isAsyncFunction from '../function/isAsync';
 import isIntegerGt from '../number/isIntegerAbove';
-import toUTC from '../date/toUTC';
-import diff from '../date/diff';
+import LRU from './LRU';
 
 /**
  * Turn a function into a memoized function. An optional resolver function can be passed which allows custom cache key generation.
@@ -11,52 +10,53 @@ import diff from '../date/diff';
  *
  * @param fn - Function to memoize
  * @param resolver - Optional resolver function to generate cache key. If not passed the first argument is used as map key
- * @param {false|number?} memoize_for - Memoize for X milliseconds, if passed false will indefinitely memoize (default = false)
+ * @param {false|number?} cache_duration_ms - Memoize for X milliseconds, if passed false will indefinitely memoize (default = false)
+ * @param {number?} cache_max_size - Memoize at max X entries, defaults to 100
  */
-function memoize <
-    T extends (...args:any[]) => unknown,
-> (fn:T, resolver?:(...args:Parameters<T>) => any, memoize_for:number|false = false):T {
-    const cache_for = isIntegerGt(memoize_for, 0) ? memoize_for : false;
+function memoize <T extends (...args:any[]) => unknown> (
+    fn:T,
+    resolver?:(...args:Parameters<T>) => any,
+    cache_duration_ms:number|false = false,
+    cache_max_size:number = 100
+):T {
+    const cache_duration = isIntegerGt(cache_duration_ms, 0) ? cache_duration_ms : false;
+    const cache = new LRU<any, {r:ReturnType<T>; ts:number}>({max_size: cache_max_size});
 
-    if (isAsyncFunction(fn)) {
-        const memoized = async function (...args:Parameters<T>) {
-            const key = typeof resolver === 'function' ? resolver(...args) : args[0];
-            if (memoized.cache.has(key)) {
-                const cached_val = memoized.cache.get(key);
-                if (cache_for === false || diff(toUTC(new Date()), cached_val.d) < cache_for) {
-                    return cached_val.r;
-                }
+    const isResolverFn = typeof resolver === 'function';
+    const memoized = isAsyncFunction(fn)
+        ? async function (...args:Parameters<T>) {
+            const key = isResolverFn ? resolver(...args) : args[0];
+            const cached_val = cache.get(key);
+
+            const now = Date.now();
+            if (cached_val !== undefined && (cache_duration === false || (now - cached_val.ts) < cache_duration)) {
+                return cached_val.r;
             }
 
-            const result = await fn(...args);
-            memoized.cache.set(key, {r: result, d: toUTC(new Date())});
+            const result = await fn(...args) as ReturnType<T>;
+            cache.set(key, {r: result, ts: now});
             return result;
-        };
-        memoized.cache = new Map();
+        }
+        : function (...args:Parameters<T>) {
+            const key = isResolverFn ? resolver(...args) : args[0];
+            const cached_val = cache.get(key);
 
-        /* eslint-disable-next-line */
-        /* @ts-ignore */
-        return memoized as T;
-    } else {
-        const memoized = function (...args:Parameters<T>) {
-            const key = typeof resolver === 'function' ? resolver(...args) : args[0];
-            if (memoized.cache.has(key)) {
-                const cached_val = memoized.cache.get(key);
-                if (cache_for === false || diff(toUTC(new Date()), cached_val.d) < cache_for) {
-                    return cached_val.r;
-                }
+            const now = Date.now();
+            if (cached_val !== undefined && (cache_duration === false || (now - cached_val.ts) < cache_duration)) {
+                return cached_val.r;
             }
 
-            const result = fn(...args);
-            memoized.cache.set(key, {r: result, d: toUTC(new Date())});
+            const result = fn(...args) as ReturnType<T>;
+            cache.set(key, {r: result, ts: now});
             return result;
         };
-        memoized.cache = new Map();
 
-        /* eslint-disable-next-line */
-        /* @ts-ignore */
-        return memoized as T;
-    }
+    /* Attach cache to memoized function for external use */
+    /* eslint-disable-next-line */
+    /* @ts-ignore */
+    memoized.cache = cache;
+
+    return memoized as T;
 }
 
 export {memoize, memoize as default};